backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
requests==2.31.0
python-decouple==3.8
pytest==7.4.4
pytest-django==4.7.0 